+++ doc/source/changelog.rst
@@ -5,6 +5,7 @@ Changelog
 Next
 ~~~~
 
+- feat: add support for adding metrics together, to evaluate files in parallel (@nryant, @Atishyy27)
 - fix: fix metrics for empty references (@Atishyy27)
 
 Version 4.1.0 (2026-05-06)